gb_bank

Module prelude

Everything needed to write banked code in one import.

Re-exports

pub use crate::bank;

pub use crate::scope;

pub use crate::Anchor;

pub use crate::Bank;

pub use crate::DynFar;

pub use crate::Far;

pub use crate::FarCall;

pub use crate::FarWith;

pub use crate::Group;

pub use crate::GroupZero;

pub use crate::Warp;

Macros

far

Take a far pointer to a banked function, for heterogeneous dispatch.

Traits

BankSafe

A value safe to carry across a bank switch: it embeds no pointer to banked code that the switch would unmap.

Attribute Macros

bank

Mark a function, static, impl block, or trait as banked.